+++ ios/Classes/GeolocationPlugin.m
@@ -1,10 +1,12 @@
-//
-//  Copyright (c) 2018 Loup Inc.
-//  Licensed under Apache License v2.0
-//
-
 #import "GeolocationPlugin.h"
+#if __has_include(<geolocation/geolocation-Swift.h>)
 #import <geolocation/geolocation-Swift.h>
+#else
+// Support project import fallback if the generated compatibility header
+// is not copied when this plugin is created as a library.
+// https://forums.swift.org/t/swift-static-libraries-dont-copy-generated-objective-c-header/19816
+#import "geolocation-Swift.h"
+#endif
 
 @implementation GeolocationPlugin
 + (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {