+++ ios/Classes/SwiftGeolocationPlugin.swift
@@ -1,8 +1,3 @@
-//
-//  Copyright (c) 2018 Loup Inc.
-//  Licensed under Apache License v2.0
-//
-
 import Flutter
 import UIKit
 import CoreLocation
@@ -10,46 +5,34 @@ import CoreLocation
 @available(iOS 9.0, *)
 public class SwiftGeolocationPlugin: NSObject, FlutterPlugin, UIApplicationDelegate {
   
-//  internal let registrar: FlutterPluginRegistrar
   private let locationClient = LocationClient()
-  private let locationChannels: LocationChannels
+  private let handler: Handler
   
   override init() {
-//    self.registrar = registrar
-    self.locationChannels = LocationChannels(locationClient: locationClient)
+    self.handler = Handler(locationClient: locationClient)
     super.init()
-
-//    registrar.addApplicationDelegate(self)
-//    locationChannels.register(on: self)
   }
   
   public static func register(with registrar: FlutterPluginRegistrar) {
-//    let channel = FlutterMethodChannel(name: "geolocation", binaryMessenger: registrar.messenger())
+    let methodChannel = FlutterMethodChannel(name: "geolocation/location", binaryMessenger: registrar.messenger())
+    let eventChannel = FlutterEventChannel(name: "geolocation/locationUpdates", binaryMessenger: registrar.messenger())
+    
     let instance = SwiftGeolocationPlugin()
-    instance.locationChannels.register(with: registrar)
+    
     registrar.addApplicationDelegate(instance)
-    
-//    registrar.addMethodCallDelegate(instance, channel: channel)
-//    registrar.addApplicationDelegate(instance)
-    
-//    _ = SwiftGeolocationPlugin(registrar: registrar)
+    registrar.addMethodCallDelegate(instance, channel: methodChannel)
+    eventChannel.setStreamHandler(instance.handler.locationUpdatesHandler)
   }
-  
-  
-  // UIApplicationDelegate
-  
-  public func applicationDidEnterBackground(_ application: UIApplication) {
-    print("DSKJHFKJHSDJF")
+
+  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+    handler.handleMethodCall(call, result: result)
   }
   
   public func applicationDidBecomeActive(_ application: UIApplication) {
-    print("BLAH!")
     locationClient.resume()
   }
   
   public func applicationWillResignActive(_ application: UIApplication) {
-    print("BLAH BLOG!")
     locationClient.pause()
   }
 }
-